+++ R/seagull-package.R
@@ -2,17 +2,17 @@
 # Copyright Michael D. Sumner
 #
 
-#' @useDynLib cgalgris
+#' @useDynLib seagull
 
-#' @name cgalgris-package
-#' @aliases cgalgris-package, cgalgris
+#' @name seagull-package
+#' @aliases seagull-package, seagull
 #' @docType package
-#' @title CGAL for gris
+#' @title CGAL for topology in R
 #' @description
 #' CGAL functionality.
 #' @details
 #' \tabular{ll}{
-#' Package: \tab cgalgris\cr
+#' Package: \tab seagull\cr
 #' Type: \tab Package\cr
 #' Version: \tab 0.0.1\cr
 #' Date: \tab 2015-05-05\cr
@@ -20,7 +20,6 @@
 #' }
 #'
 #' @author
-#' Michael D. Sumner \cr \url{http://staff.acecrc.org.au/~mdsumner} \cr \cr
 #' Maintainer: Michael D. Sumner \email{mdsumner@@gmail.com} \cr
 #' @references \url{https://github.com/mdsumner}, \url{http://www.cgal.org/}
 #' @keywords package